Refactor to add a louver model.
